NEW: chrono: add long time high resolusion timing apis
WHAT: originally we use dwt timer as get high resolusion tick source, which will overflow after UINT32_MAX/SYSCLK_FREQ(about 25.56) seconds.
Complement it with HAL tick as a low-precision clock source, now the timestamp will not get overflow for 49.71 days

Try gettick64/difftick64 apis.

TEST: test for cyclic delay 300us for 5min and timestamp works proplerly


   text	   data	    bss	    dec	    hex	filename
  37200	    560	  25400	  63160	   f6b8	build/Robocon2024-Embedded-BoardC.elf

+++ ARES/DriverAL/ares_section.h
@@ -18,6 +18,6 @@ extern unsigned int __post_os_init_start;
 extern unsigned int __post_os_init_end;
 
 extern unsigned int _etext;
-#define IS_IN_SECTION_TEXT(x) ((x) > 0 && (x) < &_etext)
+#define IS_IN_SECTION_TEXT(x) ((x) > 0 && (uint32_t)(x) < (uint32_t)&_etext)
 
 #endif
